\name{pctGCdata-package}
\alias{pctGCdata-package}
\alias{pctGCdata}
\docType{package}
\title{
  \packageTitle{pctGCdata}
}
\description{
  \packageDescription{pctGCdata}
}
\details{

This a data package that gives the GC percentage of human genome for
builds hg18 & hg19 and mouse genome for builds mm9 & mm10. The data
object, named after the build, is a list with each element giving a
vector of values for a chromosome. The GC percentage is the proportion
of nucleotides that are GC in a 1000 base window. The windows begin with
bases 1-1000 and slide 100 bases at a time.

}
\author{
\packageAuthor{pctGCdata}

Maintainer: \packageMaintainer{pctGCdata}
}
\keyword{package}
